refactored ...Dependencies/elements and ...Dependencies/index; updated NotOwnedSandboxInfo/NotOwnedSandboxInfo
