+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/NotOwnedSandboxInfo/NotOwnedSandboxInfo.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Dependencies from '../../Dependencies';
+import { Dependencies } from '../../Dependencies';
 import Files from '../../Files';
 import { Project } from '../../Project';
 import { WorkspaceItem } from '../../WorkspaceItem';
